// generated/api-client/models/event-detail-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NestJS Authentication API
 * A comprehensive NestJS authentication API with PostgreSQL & Prisma.  ### Features - JWT Authentication with Access & Refresh tokens - Google OAuth2 integration - Local authentication (email/password) - Cookie-based token storage - Protected routes with guards  ### Authentication This API uses JWT tokens stored in HTTP-only cookies for authentication.  1. Local Authentication 2. Google OAuth
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { EventOrganizerDto } from './event-organizer-dto';

/**
 * 
 * @export
 * @interface EventDetailResponseDto
 */
export interface EventDetailResponseDto {
    /**
     * 
     * @type {string}
     * @memberof EventDetailResponseDto
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof EventDetailResponseDto
     */
    'slug': string;
    /**
     * 
     * @type {string}
     * @memberof EventDetailResponseDto
     */
    'title': string;
    /**
     * 
     * @type {string}
     * @memberof EventDetailResponseDto
     */
    'type': EventDetailResponseDtoTypeEnum;
    /**
     * 
     * @type {object}
     * @memberof EventDetailResponseDto
     */
    'imageUrl': object | null;
    /**
     * 
     * @type {object}
     * @memberof EventDetailResponseDto
     */
    'description': object | null;
    /**
     * 
     * @type {object}
     * @memberof EventDetailResponseDto
     */
    'details': object | null;
    /**
     * 
     * @type {string}
     * @memberof EventDetailResponseDto
     */
    'startDate': string;
    /**
     * 
     * @type {object}
     * @memberof EventDetailResponseDto
     */
    'endDate': object | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof EventDetailResponseDto
     */
    'tags': Array<string>;
    /**
     * 
     * @type {boolean}
     * @memberof EventDetailResponseDto
     */
    'isPinned': boolean;
    /**
     * 
     * @type {string}
     * @memberof EventDetailResponseDto
     */
    'organizedBy': string;
    /**
     * 
     * @type {string}
     * @memberof EventDetailResponseDto
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof EventDetailResponseDto
     */
    'updatedAt': string;
    /**
     * 
     * @type {EventOrganizerDto}
     * @memberof EventDetailResponseDto
     */
    'organizer': EventOrganizerDto;
}

export const EventDetailResponseDtoTypeEnum = {
    IN_PERSON: 'IN_PERSON',
    VIRTUAL: 'VIRTUAL',
    HYBRID: 'HYBRID',
    OTHERS: 'OTHERS'
} as const;

export type EventDetailResponseDtoTypeEnum = typeof EventDetailResponseDtoTypeEnum[keyof typeof EventDetailResponseDtoTypeEnum];