// generated/api-client/models/event-basic-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NestJS Authentication API
 * A comprehensive NestJS authentication API with PostgreSQL & Prisma.  ### Features - JWT Authentication with Access & Refresh tokens - Google OAuth2 integration - Local authentication (email/password) - Cookie-based token storage - Protected routes with guards  ### Authentication This API uses JWT tokens stored in HTTP-only cookies for authentication.  1. Local Authentication 2. Google OAuth
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface EventBasicResponseDto
 */
export interface EventBasicResponseDto {
    /**
     * 
     * @type {string}
     * @memberof EventBasicResponseDto
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof EventBasicResponseDto
     */
    'slug': string;
    /**
     * 
     * @type {string}
     * @memberof EventBasicResponseDto
     */
    'title': string;
    /**
     * 
     * @type {string}
     * @memberof EventBasicResponseDto
     */
    'type': EventBasicResponseDtoTypeEnum;
    /**
     * 
     * @type {object}
     * @memberof EventBasicResponseDto
     */
    'imageUrl': object | null;
    /**
     * 
     * @type {object}
     * @memberof EventBasicResponseDto
     */
    'description': object | null;
    /**
     * 
     * @type {object}
     * @memberof EventBasicResponseDto
     */
    'details': object | null;
    /**
     * 
     * @type {string}
     * @memberof EventBasicResponseDto
     */
    'startDate': string;
    /**
     * 
     * @type {object}
     * @memberof EventBasicResponseDto
     */
    'endDate': object | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof EventBasicResponseDto
     */
    'tags': Array<string>;
    /**
     * 
     * @type {boolean}
     * @memberof EventBasicResponseDto
     */
    'isPinned': boolean;
    /**
     * 
     * @type {string}
     * @memberof EventBasicResponseDto
     */
    'organizedBy': string;
    /**
     * 
     * @type {string}
     * @memberof EventBasicResponseDto
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof EventBasicResponseDto
     */
    'updatedAt': string;
}

export const EventBasicResponseDtoTypeEnum = {
    IN_PERSON: 'IN_PERSON',
    VIRTUAL: 'VIRTUAL',
    HYBRID: 'HYBRID',
    OTHERS: 'OTHERS'
} as const;

export type EventBasicResponseDtoTypeEnum = typeof EventBasicResponseDtoTypeEnum[keyof typeof EventBasicResponseDtoTypeEnum];